+++ client/src/Components/ErrorPage.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const ErrorPage = () => {
+    return (
+        <div>
+            <h1>ERROR</h1>
+        </div>
+    )
+}
+
+export default ErrorPage
