+++ client/src/Components/PatientInput.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const PatientInput = () => {
+    return (
+        <div>PatientInput</div>
+    )
+}
+
+export default PatientInput
